Move Constants.java to package
