+++ src/main/java/com/modsensoftware/marketplace/controller/CategoryController.java
@@ -1,5 +1,6 @@
 package com.modsensoftware.marketplace.controller;
 
+import com.modsensoftware.marketplace.constants.Constants;
 import com.modsensoftware.marketplace.domain.Category;
 import com.modsensoftware.marketplace.dto.CategoryDto;
 import com.modsensoftware.marketplace.service.CategoryService;
