+++ src/main/java/com/modsensoftware/marketplace/controller/CompanyController.java
@@ -1,5 +1,6 @@
 package com.modsensoftware.marketplace.controller;
 
+import com.modsensoftware.marketplace.constants.Constants;
 import com.modsensoftware.marketplace.domain.Company;
 import com.modsensoftware.marketplace.dto.CompanyDto;
 import com.modsensoftware.marketplace.service.CompanyService;
